feat: add verification controls when attestion

+++ apps/main/anchor/programs/contracts/src/user.rs
@@ -1,7 +1,9 @@
 use anchor_lang::prelude::*;
 
 #[account]
+#[derive(InitSpace)]
 pub struct User {
+    #[max_len(256)]
     pub did: String,
     pub created_at: u64,
     pub last_active: u64,
@@ -14,15 +16,17 @@ pub struct User {
     pub human_verification: bool,
 
     /// points to additional data in IPFS
+    #[max_len(256)]
     pub data_cid: String,
 }
 
 #[derive(Accounts)]
+#[instruction(did: String)]
 pub struct CreateUser<'info> {
     #[account(mut)]
     pub payer: Signer<'info>,
 
-    #[account(init, payer = payer, space = 8 + User::INIT_SPACE)]
+    #[account(init, payer = payer, space = 8 + User::INIT_SPACE, seeds = [b"user", did.as_bytes()], bump)]
     pub user: Account<'info, User>,
 
     pub system_program: Program<'info, System>,
@@ -31,7 +35,7 @@ pub struct CreateUser<'info> {
 pub fn create_user(ctx: Context<CreateUser>, did: String) -> Result<()> {
     let user = &mut ctx.accounts.user;
     user.did = did;
-    user.created_at = Clock::get()?.unix_timestamp;
-    user.last_active = Clock::get()?.unix_timestamp;
+    user.created_at = Clock::get()?.unix_timestamp as u64;
+    user.last_active = Clock::get()?.unix_timestamp as u64;
     Ok(())
 }
